Simplify the form showed after adding new application
